Added coverage to the project
- Changed the compiler to use g++
- Added -Wall and fixed all given warning
- Added coverage output to document generation
- Changed Position to size_t to match its use
- Added default initializers to Match

+++ include/list.hpp
@@ -136,7 +136,7 @@ public:
 
 		std::shared_ptr<Node<T>> lp = this->root;
 
-		for (int i = 0; i < index; i++) {
+		for (size_t i = 0; i < index; i++) {
 			lp = lp->getRight();
 		}
